Move hash extension to `:core:common`
- reference to test in `StringExtensionsKtTest.kt`
