+++ core/common/src/main/java/com/shahin/core/common/extensions/StringExtensions.kt
@@ -1,4 +1,4 @@
-package com.shahin.core.database.extensions
+package com.shahin.core.common.extensions
 
 import android.util.Base64
 import java.nio.charset.StandardCharsets
